gpt-4o/requirements.txt
pandas
numpy
gradio
tqdm
huggingface_hub
torch
tqdm
openai
openai-whisper